+++ nautilus_core/model/src/data/order.rs
@@ -21,7 +21,6 @@ use std::{
 use nautilus_core::serialization::Serializable;
 use serde::{Deserialize, Serialize};
 
-use super::{quote::QuoteTick, trade::TradeTick};
 use crate::{
     enums::OrderSide,
     orderbook::{error::BookIntegrityError, ladder::BookPrice},
@@ -91,41 +90,6 @@ impl BookOrder {
             _ => panic!("{}", BookIntegrityError::NoOrderSide),
         }
     }
-
-    #[must_use]
-    pub fn from_quote_tick(tick: &QuoteTick, side: OrderSide) -> Self {
-        match side {
-            OrderSide::Buy => Self::new(
-                OrderSide::Buy,
-                tick.bid_price,
-                tick.bid_size,
-                tick.bid_price.raw as u64,
-            ),
-            OrderSide::Sell => Self::new(
-                OrderSide::Sell,
-                tick.ask_price,
-                tick.ask_size,
-                tick.ask_price.raw as u64,
-            ),
-            _ => panic!("{}", BookIntegrityError::NoOrderSide),
-        }
-    }
-
-    #[must_use]
-    pub fn from_trade_tick(tick: &TradeTick, side: OrderSide) -> Self {
-        match side {
-            OrderSide::Buy => {
-                Self::new(OrderSide::Buy, tick.price, tick.size, tick.price.raw as u64)
-            }
-            OrderSide::Sell => Self::new(
-                OrderSide::Sell,
-                tick.price,
-                tick.size,
-                tick.price.raw as u64,
-            ),
-            _ => panic!("{}", BookIntegrityError::NoOrderSide),
-        }
-    }
 }
 
 impl Default for BookOrder {
@@ -186,11 +150,7 @@ pub mod stubs {
 mod tests {
     use rstest::rstest;
 
-    use super::{stubs::*, *};
-    use crate::{
-        enums::AggressorSide,
-        identifiers::{instrument_id::InstrumentId, trade_id::TradeId},
-    };
+    use super::*;
 
     #[rstest]
     fn test_new() {
@@ -261,87 +221,5 @@ mod tests {
 
         let expected = format!("{price},{size},{side},{order_id}");
         assert_eq!(display, expected);
-    }
-
-    #[rstest]
-    #[case(OrderSide::Buy)]
-    #[case(OrderSide::Sell)]
-    fn test_from_quote_tick(#[case] side: OrderSide) {
-        let tick = QuoteTick::new(
-            InstrumentId::from("ETHUSDT-PERP.BINANCE"),
-            Price::from("5000.00"),
-            Price::from("5001.00"),
-            Quantity::from("100.000"),
-            Quantity::from("99.000"),
-            0,
-            0,
-        )
-        .unwrap();
-
-        let book_order = BookOrder::from_quote_tick(&tick, side);
-
-        assert_eq!(book_order.side, side);
-        assert_eq!(
-            book_order.price,
-            match side {
-                OrderSide::Buy => tick.bid_price,
-                OrderSide::Sell => tick.ask_price,
-                _ => panic!("Invalid test"),
-            }
-        );
-        assert_eq!(
-            book_order.size,
-            match side {
-                OrderSide::Buy => tick.bid_size,
-                OrderSide::Sell => tick.ask_size,
-                _ => panic!("Invalid test"),
-            }
-        );
-        assert_eq!(
-            book_order.order_id,
-            match side {
-                OrderSide::Buy => tick.bid_price.raw as u64,
-                OrderSide::Sell => tick.ask_price.raw as u64,
-                _ => panic!("Invalid test"),
-            }
-        );
-    }
-
-    #[rstest]
-    #[case(OrderSide::Buy)]
-    #[case(OrderSide::Sell)]
-    fn test_from_trade_tick(#[case] side: OrderSide) {
-        let tick = TradeTick::new(
-            InstrumentId::from("ETHUSDT-PERP.BINANCE"),
-            Price::from("5000.00"),
-            Quantity::from("100.00"),
-            AggressorSide::Buyer,
-            TradeId::new("1").unwrap(),
-            0,
-            0,
-        );
-
-        let book_order = BookOrder::from_trade_tick(&tick, side);
-
-        assert_eq!(book_order.side, side);
-        assert_eq!(book_order.price, tick.price);
-        assert_eq!(book_order.size, tick.size);
-        assert_eq!(book_order.order_id, tick.price.raw as u64);
-    }
-
-    #[rstest]
-    fn test_json_serialization(stub_book_order: BookOrder) {
-        let order = stub_book_order;
-        let serialized = order.as_json_bytes().unwrap();
-        let deserialized = BookOrder::from_json_bytes(serialized).unwrap();
-        assert_eq!(deserialized, order);
-    }
-
-    #[rstest]
-    fn test_msgpack_serialization(stub_book_order: BookOrder) {
-        let order = stub_book_order;
-        let serialized = order.as_msgpack_bytes().unwrap();
-        let deserialized = BookOrder::from_msgpack_bytes(serialized).unwrap();
-        assert_eq!(deserialized, order);
     }
 }