+++ nautilus_core/model/src/data/quote.rs
@@ -50,9 +50,9 @@ pub struct QuoteTick {
     pub bid_size: Quantity,
     /// The top of book ask size.
     pub ask_size: Quantity,
-    /// The UNIX timestamp (nanoseconds) when the tick event occurred.
+    /// The UNIX timestamp (nanoseconds) when the quote event occurred.
     pub ts_event: UnixNanos,
-    /// The UNIX timestamp (nanoseconds) when the data object was initialized.
+    /// The UNIX timestamp (nanoseconds) when the struct was initialized.
     pub ts_init: UnixNanos,
 }