+++ nautilus_core/model/src/orderbook/aggregation.rs
@@ -16,19 +16,26 @@
 use super::{book::OrderBook, error::InvalidBookOperation};
 use crate::{
     data::{order::BookOrder, quote::QuoteTick, trade::TradeTick},
-    enums::{BookType, OrderSide},
+    enums::{BookType, OrderSide, RecordFlag},
 };
 
-pub(crate) fn pre_process_order(book_type: BookType, mut order: BookOrder) -> BookOrder {
+pub(crate) fn pre_process_order(book_type: BookType, mut order: BookOrder, flags: u8) -> BookOrder {
     match book_type {
         BookType::L1_MBP => order.order_id = order.side as u64,
         BookType::L2_MBP => order.order_id = order.price.raw as u64,
-        BookType::L3_MBO => {} // No pre-processing
+        BookType::L3_MBO => {
+            if flags == 0 {
+            } else if RecordFlag::F_TOB.matches(flags) {
+                order.order_id = order.side as u64;
+            } else if RecordFlag::F_MBP.matches(flags) {
+                order.order_id = order.price.raw as u64;
+            }
+        }
     };
     order
 }
 
-pub fn update_book_with_quote_tick(
+pub(crate) fn update_book_with_quote_tick(
     book: &mut OrderBook,
     quote: &QuoteTick,
 ) -> Result<(), InvalidBookOperation> {
@@ -36,22 +43,27 @@ pub fn update_book_with_quote_tick(
         return Err(InvalidBookOperation::Update(book.book_type));
     };
 
-    update_book_bid(
-        book,
-        BookOrder::from_quote_tick(quote, OrderSide::Buy),
-        quote.ts_event,
-        0,
+    let bid = BookOrder::new(
+        OrderSide::Buy,
+        quote.bid_price,
+        quote.bid_size,
+        OrderSide::Buy as u64,
     );
-    update_book_ask(
-        book,
-        BookOrder::from_quote_tick(quote, OrderSide::Sell),
-        quote.ts_event,
-        0,
+
+    let ask = BookOrder::new(
+        OrderSide::Sell,
+        quote.ask_price,
+        quote.ask_size,
+        OrderSide::Sell as u64,
     );
+
+    update_book_bid(book, bid, quote.ts_event);
+    update_book_ask(book, ask, quote.ts_event);
+
     Ok(())
 }
 
-pub fn update_book_with_trade_tick(
+pub(crate) fn update_book_with_trade_tick(
     book: &mut OrderBook,
     trade: &TradeTick,
 ) -> Result<(), InvalidBookOperation> {
@@ -59,53 +71,42 @@ pub fn update_book_with_trade_tick(
         return Err(InvalidBookOperation::Update(book.book_type));
     };
 
-    update_book_bid(
-        book,
-        BookOrder::from_trade_tick(trade, OrderSide::Buy),
-        trade.ts_event,
-        0,
+    let bid = BookOrder::new(
+        OrderSide::Buy,
+        trade.price,
+        trade.size,
+        OrderSide::Buy as u64,
     );
-    update_book_ask(
-        book,
-        BookOrder::from_trade_tick(trade, OrderSide::Sell),
-        trade.ts_event,
-        0,
+
+    let ask = BookOrder::new(
+        OrderSide::Sell,
+        trade.price,
+        trade.size,
+        OrderSide::Sell as u64,
     );
+
+    update_book_bid(book, bid, trade.ts_event);
+    update_book_ask(book, ask, trade.ts_event);
+
     Ok(())
 }
 
-pub fn update_book_ask(book: &mut OrderBook, order: BookOrder, ts_event: u64, sequence: u64) {
-    match book.asks.top() {
-        Some(top_asks) => match top_asks.first() {
-            Some(top_ask) => {
-                let order_id = top_ask.order_id;
-                book.asks.remove(order_id, ts_event, sequence);
-                book.asks.add(order);
-            }
-            None => {
-                book.asks.add(order);
-            }
-        },
-        None => {
-            book.asks.add(order);
+fn update_book_ask(book: &mut OrderBook, order: BookOrder, ts_event: u64) {
+    if let Some(top_asks) = book.asks.top() {
+        if let Some(top_ask) = top_asks.first() {
+            let order_id = top_ask.order_id;
+            book.asks.remove(order_id, 0, ts_event);
         }
     }
+    book.asks.add(order);
 }
 
-pub fn update_book_bid(book: &mut OrderBook, order: BookOrder, ts_event: u64, sequence: u64) {
-    match book.bids.top() {
-        Some(top_bids) => match top_bids.first() {
-            Some(top_bid) => {
-                let order_id = top_bid.order_id;
-                book.bids.remove(order_id, ts_event, sequence);
-                book.bids.add(order);
-            }
-            None => {
-                book.bids.add(order);
-            }
-        },
-        None => {
-            book.bids.add(order);
+fn update_book_bid(book: &mut OrderBook, order: BookOrder, ts_event: u64) {
+    if let Some(top_bids) = book.bids.top() {
+        if let Some(top_bid) = top_bids.first() {
+            let order_id = top_bid.order_id;
+            book.bids.remove(order_id, 0, ts_event);
         }
     }
+    book.bids.add(order);
 }