+++ nautilus_core/model/src/orderbook/level.rs
@@ -143,11 +143,11 @@ impl Level {
         self.update_insertion_order();
     }
 
-    pub fn remove_by_id(&mut self, order_id: OrderId, ts_event: u64, sequence: u64) {
+    pub fn remove_by_id(&mut self, order_id: OrderId, sequence: u64, ts_event: u64) {
         assert!(
             self.orders.remove(&order_id).is_some(),
             "{}",
-            &BookIntegrityError::OrderNotFound(order_id, ts_event, sequence)
+            &BookIntegrityError::OrderNotFound(order_id, sequence, ts_event)
         );
         self.update_insertion_order();
     }
@@ -365,7 +365,7 @@ mod tests {
 
     #[rstest]
     #[should_panic(
-        expected = "Integrity error: order not found: order_id=1, ts_event=2, sequence=3"
+        expected = "Integrity error: order not found: order_id=1, sequence=2, ts_event=3"
     )]
     fn test_remove_nonexistent_order() {
         let mut level = Level::new(BookPrice::new(Price::from("1.00"), OrderSide::Buy));